services/regulatory_compliance/requirements.txt
google-cloud-pubsub==2.13.0
google-cloud-bigquery==3.18.0